[Refactor] Allow JSON serializable objects in expected values
This refactor changes the assertions so that they correctly JSON
encode and then decode the expected values before comparing them
with the decoded document value. This allows the developer to pass
JSON serializable objects within the expected values without having
to manually convert them to the expected JSON value. It also fixes
a bug with comparison of float and integer values.

This refactor also sorts out method type-hinting.